Refactor to clean up access
