+++ orchagent/dash/dashorch.cpp
@@ -1295,30 +1295,6 @@ void DashOrch::doTask(ConsumerBase& consumer)
     }
 }
 
-template<>
-void DashOrch::DashCounter<CounterType::ENI>::fetchStats()
-{
-    counter_stats.clear();
-    auto stat_enum_list = queryAvailableCounterStats((sai_object_type_t)SAI_OBJECT_TYPE_ENI);
-    for (auto &stat_enum: stat_enum_list)
-    {
-        auto counter_id = static_cast<sai_eni_stat_t>(stat_enum);
-        counter_stats.insert(sai_serialize_eni_stat(counter_id));
-    }
-}
-
-template<>
-void DashOrch::DashCounter<CounterType::DASH_METER>::fetchStats()
-{
-    counter_stats.clear();
-    auto stat_enum_list = queryAvailableCounterStats((sai_object_type_t)SAI_OBJECT_TYPE_METER_BUCKET_ENTRY);
-    for (auto &stat_enum: stat_enum_list)
-    {
-        auto counter_id = static_cast<sai_meter_bucket_entry_stat_t>(stat_enum);
-        counter_stats.insert(sai_serialize_meter_bucket_entry_stat(counter_id));
-    }
-}
-
 void DashOrch::addEniMapEntry(sai_object_id_t oid, const string &name) 
 {
     SWSS_LOG_ENTER();
